dvc: split remotes from trees
Related to #4050
