+++ dvc/dependency/http.py
@@ -1,6 +1,7 @@
 from dvc.dependency.base import BaseDependency
 from dvc.output.base import BaseOutput
-from dvc.remote.http import HTTPRemoteTree
+
+from ..tree.http import HTTPRemoteTree
 
 
 class HTTPDependency(BaseDependency, BaseOutput):
