+++ dvc/output/gs.py
@@ -1,5 +1,6 @@
 from dvc.output.s3 import S3Output
-from dvc.remote.gs import GSRemoteTree
+
+from ..tree.gs import GSRemoteTree
 
 
 class GSOutput(S3Output):
